extracted inner App's parent component
