+++ client/src/components/dashboard.js
@@ -1,38 +1,22 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Link } from 'react-router-dom';
-import Drawer from './drawer';
-import Navigation from './navigation';
 
 import './dashboard.css';
 
-function Dashboard({ channels }) {
-  const [open, setOpen] = useState(false);
-  const toggleDrawer = () => {
-    setOpen(!open);
-  };
-
+function Dashboard() {
   return (
-    // TODO: extract parent component and only leave dashboard
-    <div className="parent">
-      <Drawer channels={channels} toggleDrawer={toggleDrawer} isOpen={open} />
-
-      <div className="tabContent">
-        <Navigation toggleDrawer={toggleDrawer} />
-
-        <div className="dashboard">
-          <p>Welcome to your dashboard ! Go ahead and...</p>
-          <div>
-            <input
-              className="searchInput"
-              type="text"
-              placeholder="Find existing channels"
-            />{' '}
-            <span className="actionsDivider">or...</span>{' '}
-            <Link className="createLink" to="/channels/new">
-              Create a new one
-            </Link>
-          </div>
-        </div>
+    <div className="dashboard">
+      <p>Welcome to your dashboard ! Go ahead and...</p>
+      <div>
+        <input
+          className="searchInput"
+          type="text"
+          placeholder="Find existing channels"
+        />{' '}
+        <span className="actionsDivider">or...</span>{' '}
+        <Link className="createLink" to="/channels/new">
+          Create a new one
+        </Link>
       </div>
     </div>
   );
